Add files via upload
1. Assuming that a web server with the architecture spends 20% on processing,
30% on disk access, and 50% on network transfer, you have a base system
consisting of a 500MHz processor and a disk with 20Mbytes/sec data transfer
rate. This system costs $5K and can support 10,000 average web page
accesses/sec. Considering the following three options to enhance system
performance:
a. Option-1: replacing the existing disk with a disk supporting
40Mbytes/sec data transfer rate with an additional (compared to the
base) cost of $1,000.
b. Option-2: replacing the processor with an 800MHz processor with an
additional (compared to the base) cost of $800.
c. Option-3: using the two enhancements indicated in Option-1 &
Option-2 together with an additional (compared to the base) cost of
$1,500.
(1) Determine what will be the new performance (in terms of average web
page access per second) with each of the enhancement options.
(2) By doing a cost-performance analysis, determine which option will be
cost-effective to go for and why.

2. The Amdahl’s law is based on the assumption that when an enhancement is
performed to some part of the system, the enhancement doesn’t have any
negative impact on the non-enhanced part. However, in real life, it could lead
to negative impact on these parts. Thus, Amdahl’s law can be modified to take
care of this situation.
Consider a computer system with two components, A and B, which can be
enhanced. There is interdependency between these components. And
enhancement in one component affects the other. There are three options for
enhancement as suggested below. All options involve the same amount of cost.
a. Option-A: Let us assume that 𝒇𝑨, the fraction of instructions using
component A, can be sped up by 10 times. However, due to the
dependency of A on B, another fraction 2𝒇𝑨 will be get slowed down by
5 times.
b. Option-B: The instructions using component B, fraction 𝒇𝑩, can be
sped up by 20 times. The dependency forces another fraction 0.5𝒇𝑩 to
get slow down by 2 times.
c. Option-C: A fraction 𝒇𝑨 of instructions using the component A, can be
sped up by a factor of 4. Unfortunately, the dependency forces another
fraction 𝒇𝑨 to get slowed down by 1.8 times.
(1) Derive the parameterized speed-up equations (in terms of 𝒇𝑨, 𝒇𝑩) for
each of the above three options.
(2) As a beginner architect, which option will be preferred and why? Give
convincing reasoning. Assuming for a reasonable enhancement, you
need to have 𝒇𝑨 > 𝟎 and 𝒇𝑩 > 𝟎

3. A set of three systems are being evaluated to be used in a laboratory
environment. This environment uses three types of programs with a relative
usage of 45% (Program 1), 35% (Program 2), and 20% (Program 3)
respectively. Each of these three programs has been benchmarked on these
three systems individually and their execution times are shown as follows.
Programs
System 1
System 2
System 3
Programs 1
1.0 sec
2.0 sec
1.5 sec
Programs 2
10.0 sec
7.0 sec
5.0 sec
Programs 3
5.0 sec
3.0 sec
4.0 sec
a. Determine which of the above three systems will provide the best
performance for the laboratory.
b. The three systems cost as follows: $8,000 (System 1), $5,000 (System 2),
and $6,500 (System 3). By doing a cost-performance analysis, indicate
which one of these systems you will choose and why.